* Profile Model Bean generated.
